backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
django-environ==0.11.2
Pillow==10.1.0
pyzk==0.9
requests==2.32.5
pytz==2024.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
openpyxl==3.1.2

